DTSCCI-685 extract logic
