+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/respondtoclaimspeccallbackhandlertaskstests/HandleDefendAllClaim.java
@@ -21,6 +21,7 @@ import uk.gov.hmcts.reform.civil.utils.MonetaryConversions;
 import uk.gov.hmcts.reform.civil.validation.PaymentDateValidator;
 
 import java.math.BigDecimal;
+import java.util.EnumSet;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
@@ -32,13 +33,18 @@ import static uk.gov.hmcts.reform.civil.enums.RespondentResponseTypeSpec.FULL_DE
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.NO;
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.YES;
 import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.BOTH_RESPONDENTS_DISPUTE;
-import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.CAN_ANSWER_RESPONDENT_1;
-import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.CAN_ANSWER_RESPONDENT_2;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.NEED_FINANCIAL_DETAILS_1;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.NEED_FINANCIAL_DETAILS_2;
 import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.ONLY_RESPONDENT_1_DISPUTES;
 import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.REPAYMENT_PLAN_2;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.RESPONDENT_1_ADMITS_PART_OR_FULL;
 import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.RESPONDENT_1_PAID_LESS;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.RESPONDENT_2_ADMITS_PART_OR_FULL;
 import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.RESPONDENT_2_PAID_LESS;
 import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.SOMEONE_DISPUTES;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.WHEN_WILL_CLAIM_BE_PAID;
+import static uk.gov.hmcts.reform.civil.handler.callback.user.spec.show.DefendantResponseShowTag.WHY_2_DOES_NOT_PAY_IMMEDIATELY;
 
 @Component
 @RequiredArgsConstructor
@@ -48,217 +54,300 @@ public class HandleDefendAllClaim implements CaseTask {
     private final ObjectMapper objectMapper;
     private final FeatureToggleService toggleService;
     private final PaymentDateValidator paymentDateValidator;
-    private final RespondToClaimSpecUtils respondToClaimSpecUtilsDisputeDetails;
 
     static final String UNKNOWN_MP_SCENARIO = "Unknown mp scenario";
 
     public CallbackResponse execute(CallbackParams callbackParams) {
-        log.info("Executing HandleDefendAllClaim with callbackParams: {}", callbackParams);
         CaseData caseData = callbackParams.getCaseData();
-        List<String> errors = validatePayments(caseData);
+        List<String> errors = paymentDateValidator.validate(Optional.ofNullable(caseData.getRespondToClaim())
+                                                                .orElseGet(() -> RespondToClaim.builder().build()));
         if (!errors.isEmpty()) {
-            log.error("Validation errors found: {}", errors);
-            return buildErrorResponse(errors);
+            return AboutToStartOrSubmitCallbackResponse.builder()
+                .errors(errors)
+                .build();
         }
-
         CaseData.CaseDataBuilder<?, ?> updatedCase = caseData.toBuilder();
         updatedCase.showConditionFlags(whoDisputesFullDefence(caseData));
-
         if (SpecJourneyConstantLRSpec.DEFENDANT_RESPONSE_SPEC.equals(callbackParams.getRequest().getEventId())) {
-            log.info("Populating respondent response type spec paid status");
             populateRespondentResponseTypeSpecPaidStatus(caseData, updatedCase);
-            updateSpecPaidOrDisputeStatus(caseData, updatedCase);
-            updatedCase.responseClaimTrack(getAllocatedTrack(caseData).name());
+            if (caseData.getRespondent1ClaimResponsePaymentAdmissionForSpec()
+                == RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT
+                || DISPUTES_THE_CLAIM.equals(caseData.getDefenceRouteRequired())
+                || DISPUTES_THE_CLAIM.equals(caseData.getDefenceRouteRequired2())
+                || caseData.getRespondent1ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION
+                || caseData.getRespondent2ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION) {
+                updatedCase.specPaidLessAmountOrDisputesOrPartAdmission(YES);
+            } else {
+                updatedCase.specPaidLessAmountOrDisputesOrPartAdmission(NO);
+            }
+            if (YES.equals(caseData.getIsRespondent2())) {
+                if (RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT
+                    != caseData.getRespondent1ClaimResponsePaymentAdmissionForSpec()
+                    && (DISPUTES_THE_CLAIM.equals(caseData.getDefenceRouteRequired2())
+                    || caseData.getRespondent2ClaimResponseTypeForSpec()
+                    == RespondentResponseTypeSpec.PART_ADMISSION)) {
+                    updatedCase.specDisputesOrPartAdmission(YES);
+                } else {
+                    updatedCase.specDisputesOrPartAdmission(NO);
+                }
+            } else {
+                if (RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT
+                    != caseData.getRespondent1ClaimResponsePaymentAdmissionForSpec()
+                    && (DISPUTES_THE_CLAIM.equals(caseData.getDefenceRouteRequired())
+                    || caseData.getRespondent1ClaimResponseTypeForSpec()
+                    == RespondentResponseTypeSpec.PART_ADMISSION)) {
+                    updatedCase.specDisputesOrPartAdmission(YES);
+                } else {
+                    updatedCase.specDisputesOrPartAdmission(NO);
+                }
+            }
+            AllocatedTrack allocatedTrack = getAllocatedTrack(caseData);
+            updatedCase.responseClaimTrack(allocatedTrack.name());
         }
-
-        log.info("Building callback response");
-        return buildCallbackResponse(updatedCase);
-    }
-
-    private List<String> validatePayments(CaseData caseData) {
-        log.info("Validating payments for caseData: {}", caseData);
-        return paymentDateValidator.validate(Optional.ofNullable(caseData.getRespondToClaim())
-                                                 .orElseGet(() -> RespondToClaim.builder().build()));
-    }
-
-    private CallbackResponse buildErrorResponse(List<String> errors) {
-        log.error("Building error response with errors: {}", errors);
-        return AboutToStartOrSubmitCallbackResponse.builder().errors(errors).build();
-    }
-
-    private void updateSpecPaidOrDisputeStatus(CaseData caseData, CaseData.CaseDataBuilder<?, ?> updatedCase) {
-        log.info("Updating spec paid or dispute status for caseData: {}", caseData);
-        if (isPaidLessOrDispute(caseData)) {
-            updatedCase.specPaidLessAmountOrDisputesOrPartAdmission(YES);
-        } else {
-            updatedCase.specPaidLessAmountOrDisputesOrPartAdmission(NO);
-        }
-
-        if (isSpecDisputeOrPartAdmission(caseData)) {
-            updatedCase.specDisputesOrPartAdmission(YES);
-        } else {
-            updatedCase.specDisputesOrPartAdmission(NO);
-        }
-    }
-
-    private boolean isPaidLessOrDispute(CaseData caseData) {
-        return isPaidLessForRespondent1(caseData)
-            || isDisputeForRespondent(caseData.getDefenceRouteRequired())
-            || isDisputeForRespondent(caseData.getDefenceRouteRequired2())
-            || isPartAdmissionForRespondent(caseData.getRespondent1ClaimResponseTypeForSpec())
-            || isPartAdmissionForRespondent(caseData.getRespondent2ClaimResponseTypeForSpec());
-    }
-
-    private boolean isPaidLessForRespondent1(CaseData caseData) {
-        return caseData.getRespondent1ClaimResponsePaymentAdmissionForSpec() == RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT;
-    }
-
-    private boolean isDisputeForRespondent(String defenceRoute) {
-        return DISPUTES_THE_CLAIM.equals(defenceRoute);
-    }
-
-    private boolean isPartAdmissionForRespondent(RespondentResponseTypeSpec responseType) {
-        return responseType == RespondentResponseTypeSpec.PART_ADMISSION;
-    }
-
-    private boolean isSpecDisputeOrPartAdmission(CaseData caseData) {
-        return YES.equals(caseData.getIsRespondent2())
-            ? isSpecDisputeOrPartAdmissionForRespondent(caseData.getDefenceRouteRequired2(),
-                                                        caseData.getRespondent2ClaimResponseTypeForSpec(), caseData.getRespondent1ClaimResponsePaymentAdmissionForSpec())
-            : isSpecDisputeOrPartAdmissionForRespondent(caseData.getDefenceRouteRequired(),
-                                                        caseData.getRespondent1ClaimResponseTypeForSpec(), caseData.getRespondent1ClaimResponsePaymentAdmissionForSpec());
-    }
-
-    private boolean isSpecDisputeOrPartAdmissionForRespondent(
-        String defenceRoute, RespondentResponseTypeSpec responseType, RespondentResponseTypeSpecPaidStatus paymentStatus) {
-
-        return RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT != paymentStatus
-            && (DISPUTES_THE_CLAIM.equals(defenceRoute) || responseType == RespondentResponseTypeSpec.PART_ADMISSION);
+        return AboutToStartOrSubmitCallbackResponse.builder()
+            .data(updatedCase.build().toMap(objectMapper))
+            .build();
     }
 
     private Set<DefendantResponseShowTag> whoDisputesFullDefence(CaseData caseData) {
-        log.info("Determining who disputes full defence for caseData: {}", caseData);
         Set<DefendantResponseShowTag> tags = new HashSet<>(caseData.getShowConditionFlags());
-        respondToClaimSpecUtilsDisputeDetails.removeWhoDisputesAndWhoPaidLess(tags);
-
+        // in case of backtracking
+        removeWhoDisputesAndWhoPaidLess(tags);
+        Set<DefendantResponseShowTag> bcoPartAdmission = whoDisputesBcoPartAdmission(caseData);
         MultiPartyScenario mpScenario = getMultiPartyScenario(caseData);
-        Set<DefendantResponseShowTag> bcoPartAdmission = respondToClaimSpecUtilsDisputeDetails.whoDisputesBcoPartAdmission(caseData);
-
-        handleScenario(mpScenario, caseData, tags, bcoPartAdmission);
-        addSomeoneDisputesTag(tags);
-        return tags;
-    }
-
-    private void handleScenario(MultiPartyScenario mpScenario, CaseData caseData,
-                                Set<DefendantResponseShowTag> tags, Set<DefendantResponseShowTag> bcoPartAdmission) {
-        log.info("Handling scenario: {} for caseData: {}", mpScenario, caseData);
         switch (mpScenario) {
             case ONE_V_ONE:
-                handleOneVOneScenario(caseData, bcoPartAdmission);
+                fullDefenceAndPaidLess(
+                    caseData.getRespondent1ClaimResponseTypeForSpec(),
+                    caseData.getDefenceRouteRequired(),
+                    caseData.getRespondToClaim(),
+                    caseData.getTotalClaimAmount(),
+                    ONLY_RESPONDENT_1_DISPUTES,
+                    DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                ).ifPresent(bcoPartAdmission::add);
                 break;
             case TWO_V_ONE:
-                handleTwoVOneScenario(caseData, bcoPartAdmission);
+                if (!bcoPartAdmission.contains(ONLY_RESPONDENT_1_DISPUTES)) {
+                    if (caseData.getDefendantSingleResponseToBothClaimants() == YES) {
+                        fullDefenceAndPaidLess(
+                            caseData.getRespondent1ClaimResponseTypeForSpec(),
+                            caseData.getDefenceRouteRequired(),
+                            caseData.getRespondToClaim(),
+                            caseData.getTotalClaimAmount(),
+                            ONLY_RESPONDENT_1_DISPUTES,
+                            DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                        ).ifPresent(bcoPartAdmission::add);
+                    } else {
+                        fullDefenceAndPaidLess(
+                            caseData.getClaimant1ClaimResponseTypeForSpec(),
+                            caseData.getDefenceRouteRequired(),
+                            caseData.getRespondToClaim(),
+                            caseData.getTotalClaimAmount(),
+                            ONLY_RESPONDENT_1_DISPUTES,
+                            DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                        ).ifPresent(bcoPartAdmission::add);
+
+                        fullDefenceAndPaidLess(
+                            caseData.getClaimant2ClaimResponseTypeForSpec(),
+                            caseData.getDefenceRouteRequired(),
+                            caseData.getRespondToClaim(),
+                            caseData.getTotalClaimAmount(),
+                            ONLY_RESPONDENT_1_DISPUTES,
+                            DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                        ).ifPresent(bcoPartAdmission::add);
+                    }
+                }
                 break;
             case ONE_V_TWO_ONE_LEGAL_REP:
-                handleOneVTwoOneLegalRepScenario(caseData, bcoPartAdmission);
+                if (caseData.getRespondentResponseIsSame() == YES) {
+                    fullDefenceAndPaidLess(
+                        caseData.getRespondent1ClaimResponseTypeForSpec(),
+                        caseData.getDefenceRouteRequired(),
+                        caseData.getRespondToClaim(),
+                        caseData.getTotalClaimAmount(),
+                        BOTH_RESPONDENTS_DISPUTE,
+                        DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                    ).ifPresent(tag -> {
+                        bcoPartAdmission.add(tag);
+                        if (tag == DefendantResponseShowTag.RESPONDENT_1_PAID_LESS) {
+                            bcoPartAdmission.add(DefendantResponseShowTag.RESPONDENT_2_PAID_LESS);
+                        }
+                    });
+                } else {
+                    fullDefenceAndPaidLess(
+                        caseData.getRespondent1ClaimResponseTypeForSpec(),
+                        caseData.getDefenceRouteRequired(),
+                        caseData.getRespondToClaim(),
+                        caseData.getTotalClaimAmount(),
+                        ONLY_RESPONDENT_1_DISPUTES,
+                        DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                    ).ifPresent(bcoPartAdmission::add);
+                    if (caseData.getRespondentResponseIsSame() == YES) {
+                        if (bcoPartAdmission.contains(RESPONDENT_1_PAID_LESS)) {
+                            bcoPartAdmission.add(RESPONDENT_2_PAID_LESS);
+                        }
+                    } else {
+                        fullDefenceAndPaidLess(
+                            caseData.getRespondent2ClaimResponseTypeForSpec(),
+                            // if only 2nd defends, defenceRouteRequired2 field is not used
+                            caseData.getDefenceRouteRequired(),
+                            caseData.getRespondToClaim(),
+                            caseData.getTotalClaimAmount(),
+                            DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES,
+                            DefendantResponseShowTag.RESPONDENT_2_PAID_LESS
+                        ).ifPresent(bcoPartAdmission::add);
+                    }
+                    EnumSet<DefendantResponseShowTag> bothOnlyDisputes = EnumSet.of(
+                        DefendantResponseShowTag.ONLY_RESPONDENT_1_DISPUTES,
+                        DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES
+                    );
+                    if (bcoPartAdmission.containsAll(bothOnlyDisputes)) {
+                        bcoPartAdmission.removeAll(bothOnlyDisputes);
+                        bcoPartAdmission.add(BOTH_RESPONDENTS_DISPUTE);
+                    }
+                }
                 break;
             case ONE_V_TWO_TWO_LEGAL_REP:
-                handleOneVTwoTwoLegalRepScenario(caseData, tags, bcoPartAdmission);
+                if (tags.contains(DefendantResponseShowTag.CAN_ANSWER_RESPONDENT_1)) {
+                    fullDefenceAndPaidLess(
+                        caseData.getRespondent1ClaimResponseTypeForSpec(),
+                        caseData.getDefenceRouteRequired(),
+                        caseData.getRespondToClaim(),
+                        caseData.getTotalClaimAmount(),
+                        ONLY_RESPONDENT_1_DISPUTES,
+                        DefendantResponseShowTag.RESPONDENT_1_PAID_LESS
+                    ).ifPresent(bcoPartAdmission::add);
+                } else if (tags.contains(DefendantResponseShowTag.CAN_ANSWER_RESPONDENT_2)) {
+                    fullDefenceAndPaidLess(
+                        caseData.getRespondent2ClaimResponseTypeForSpec(),
+                        caseData.getDefenceRouteRequired2(),
+                        caseData.getRespondToClaim2(),
+                        caseData.getTotalClaimAmount(),
+                        DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES,
+                        DefendantResponseShowTag.RESPONDENT_2_PAID_LESS
+                    ).ifPresent(bcoPartAdmission::add);
+                }
                 break;
             default:
-                log.error("Unknown multi-party scenario: {}", mpScenario);
                 throw new UnsupportedOperationException(UNKNOWN_MP_SCENARIO);
         }
-    }
-
-    private void handleOneVOneScenario(CaseData caseData, Set<DefendantResponseShowTag> tags) {
-        log.info("Handling ONE_V_ONE scenario for caseData: {}", caseData);
-        fullDefenceAndPaidLess(
-            caseData.getRespondent1ClaimResponseTypeForSpec(),
-            caseData.getDefenceRouteRequired(),
-            caseData.getRespondToClaim(),
-            caseData.getTotalClaimAmount(),
-            ONLY_RESPONDENT_1_DISPUTES,
-            RESPONDENT_1_PAID_LESS
-        ).ifPresent(tags::add);
-    }
-
-    private void handleTwoVOneScenario(CaseData caseData, Set<DefendantResponseShowTag> tags) {
-        log.info("Handling TWO_V_ONE scenario for caseData: {}", caseData);
-        if (YES.equals(caseData.getDefendantSingleResponseToBothClaimants())) {
-            handleOneVOneScenario(caseData, tags);
-        } else {
-            fullDefenceAndPaidLess(
-                caseData.getClaimant1ClaimResponseTypeForSpec(),
-                caseData.getDefenceRouteRequired(),
-                caseData.getRespondToClaim(),
-                caseData.getTotalClaimAmount(),
-                ONLY_RESPONDENT_1_DISPUTES,
-                RESPONDENT_1_PAID_LESS
-            ).ifPresent(tags::add);
-
-            fullDefenceAndPaidLess(
-                caseData.getClaimant2ClaimResponseTypeForSpec(),
-                caseData.getDefenceRouteRequired(),
-                caseData.getRespondToClaim(),
-                caseData.getTotalClaimAmount(),
-                ONLY_RESPONDENT_1_DISPUTES,
-                RESPONDENT_1_PAID_LESS
-            ).ifPresent(tags::add);
-        }
-    }
-
-    private void handleOneVTwoOneLegalRepScenario(CaseData caseData, Set<DefendantResponseShowTag> tags) {
-        log.info("Handling ONE_V_TWO_ONE_LEGAL_REP scenario for caseData: {}", caseData);
-        if (YES.equals(caseData.getRespondentResponseIsSame())) {
-            fullDefenceAndPaidLess(
-                caseData.getRespondent1ClaimResponseTypeForSpec(),
-                caseData.getDefenceRouteRequired(),
-                caseData.getRespondToClaim(),
-                caseData.getTotalClaimAmount(),
-                BOTH_RESPONDENTS_DISPUTE,
-                RESPONDENT_1_PAID_LESS
-            ).ifPresent(tag -> {
-                tags.add(tag);
-                if (tag == RESPONDENT_1_PAID_LESS) {
-                    tags.add(RESPONDENT_2_PAID_LESS);
-                }
-            });
-        } else {
-            handleOneVOneScenario(caseData, tags);
-        }
-    }
-
-    private void handleOneVTwoTwoLegalRepScenario(CaseData caseData, Set<DefendantResponseShowTag> tags, Set<DefendantResponseShowTag> bcoPartAdmission) {
-        log.info("Handling ONE_V_TWO_TWO_LEGAL_REP scenario for caseData: {}", caseData);
-        if (tags.contains(CAN_ANSWER_RESPONDENT_1)) {
-            fullDefenceAndPaidLess(
-                caseData.getRespondent1ClaimResponseTypeForSpec(),
-                caseData.getDefenceRouteRequired(),
-                caseData.getRespondToClaim(),
-                caseData.getTotalClaimAmount(),
-                ONLY_RESPONDENT_1_DISPUTES,
-                RESPONDENT_1_PAID_LESS
-            ).ifPresent(bcoPartAdmission::add);
-        } else if (tags.contains(CAN_ANSWER_RESPONDENT_2)) {
-            fullDefenceAndPaidLess(
-                caseData.getRespondent2ClaimResponseTypeForSpec(),
-                caseData.getDefenceRouteRequired2(),
-                caseData.getRespondToClaim2(),
-                caseData.getTotalClaimAmount(),
-                ONLY_RESPONDENT_2_DISPUTES,
-                RESPONDENT_2_PAID_LESS
-            ).ifPresent(bcoPartAdmission::add);
-        }
-    }
-
-    private void addSomeoneDisputesTag(Set<DefendantResponseShowTag> tags) {
-        log.info("Adding SOMEONE_DISPUTES tag if applicable");
+        tags.addAll(bcoPartAdmission);
         if (tags.contains(ONLY_RESPONDENT_1_DISPUTES)
             || tags.contains(ONLY_RESPONDENT_2_DISPUTES)
             || tags.contains(BOTH_RESPONDENTS_DISPUTE)) {
             tags.add(SOMEONE_DISPUTES);
         }
+        return tags;
+    }
+
+    private void populateRespondentResponseTypeSpecPaidStatus(CaseData caseData,
+                                                              CaseData.CaseDataBuilder<?, ?> updated) {
+        if (SpecJourneyConstantLRSpec.HAS_PAID_THE_AMOUNT_CLAIMED.equals(caseData.getDefenceRouteRequired())
+            && caseData.getRespondToClaim().getHowMuchWasPaid() != null) {
+            // CIV-208 howMuchWasPaid is pence, totalClaimAmount is pounds, hence the need for conversion
+            int comparison = caseData.getRespondToClaim().getHowMuchWasPaid()
+                .compareTo(new BigDecimal(MonetaryConversions.poundsToPennies(caseData.getTotalClaimAmount())));
+            if (comparison < 0) {
+                updated.respondent1ClaimResponsePaymentAdmissionForSpec(
+                    RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT).build();
+            } else {
+                updated.respondent1ClaimResponsePaymentAdmissionForSpec(
+                    RespondentResponseTypeSpecPaidStatus.PAID_FULL_OR_MORE_THAN_CLAIMED_AMOUNT).build();
+            }
+        } else {
+            updated.respondent1ClaimResponsePaymentAdmissionForSpec(RespondentResponseTypeSpecPaidStatus.DID_NOT_PAY)
+                .build();
+        }
+
+        if (YES.equals(caseData.getIsRespondent2())) {
+            if (SpecJourneyConstantLRSpec.HAS_PAID_THE_AMOUNT_CLAIMED.equals(caseData.getDefenceRouteRequired2())
+                && caseData.getRespondToClaim2().getHowMuchWasPaid() != null) {
+                // CIV-208 howMuchWasPaid is pence, totalClaimAmount is pounds, hence the need for conversion
+                int comparison = caseData.getRespondToClaim2().getHowMuchWasPaid()
+                    .compareTo(new BigDecimal(MonetaryConversions.poundsToPennies(caseData.getTotalClaimAmount())));
+                if (comparison < 0) {
+                    updated.respondent1ClaimResponsePaymentAdmissionForSpec(
+                        RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT).build();
+                } else {
+                    updated.respondent1ClaimResponsePaymentAdmissionForSpec(
+                        RespondentResponseTypeSpecPaidStatus.PAID_FULL_OR_MORE_THAN_CLAIMED_AMOUNT).build();
+                }
+            } else {
+                updated.respondent1ClaimResponsePaymentAdmissionForSpec(null).build();
+            }
+        }
+    }
+
+    private AllocatedTrack getAllocatedTrack(CaseData caseData) {
+        return AllocatedTrack.getAllocatedTrack(caseData.getTotalClaimAmount(), null, null,
+                                                toggleService, caseData);
+    }
+
+    private void removeWhoDisputesAndWhoPaidLess(Set<DefendantResponseShowTag> tags) {
+        tags.removeIf(EnumSet.of(
+            ONLY_RESPONDENT_1_DISPUTES,
+            DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES,
+            DefendantResponseShowTag.BOTH_RESPONDENTS_DISPUTE,
+            SOMEONE_DISPUTES,
+            DefendantResponseShowTag.CURRENT_ADMITS_PART_OR_FULL,
+            DefendantResponseShowTag.RESPONDENT_1_PAID_LESS,
+            DefendantResponseShowTag.RESPONDENT_2_PAID_LESS,
+            WHEN_WILL_CLAIM_BE_PAID,
+            RESPONDENT_1_ADMITS_PART_OR_FULL,
+            RESPONDENT_2_ADMITS_PART_OR_FULL,
+            NEED_FINANCIAL_DETAILS_1,
+            NEED_FINANCIAL_DETAILS_2,
+            DefendantResponseShowTag.WHY_1_DOES_NOT_PAY_IMMEDIATELY,
+            WHY_2_DOES_NOT_PAY_IMMEDIATELY,
+            REPAYMENT_PLAN_2,
+            DefendantResponseShowTag.MEDIATION
+        )::contains);
+    }
+
+    private Set<DefendantResponseShowTag> whoDisputesBcoPartAdmission(CaseData caseData) {
+        Set<DefendantResponseShowTag> tags = EnumSet.noneOf(DefendantResponseShowTag.class);
+        MultiPartyScenario mpScenario = getMultiPartyScenario(caseData);
+        switch (mpScenario) {
+            case ONE_V_ONE:
+                if (caseData.getRespondent1ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION) {
+                    tags.add(ONLY_RESPONDENT_1_DISPUTES);
+                }
+                break;
+            case TWO_V_ONE:
+                if ((caseData.getDefendantSingleResponseToBothClaimants() == YES
+                    && caseData.getRespondent1ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION)
+                    || caseData.getClaimant1ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION
+                    || caseData.getClaimant2ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION) {
+                    tags.add(ONLY_RESPONDENT_1_DISPUTES);
+                }
+                break;
+            case ONE_V_TWO_ONE_LEGAL_REP:
+                if (caseData.getRespondent1ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION) {
+                    if (caseData.getRespondentResponseIsSame() == YES
+                        || caseData.getRespondent2ClaimResponseTypeForSpec()
+                        == RespondentResponseTypeSpec.PART_ADMISSION) {
+                        tags.add(DefendantResponseShowTag.BOTH_RESPONDENTS_DISPUTE);
+                    } else {
+                        tags.add(ONLY_RESPONDENT_1_DISPUTES);
+                    }
+                } else if (caseData.getRespondent2ClaimResponseTypeForSpec()
+                    == RespondentResponseTypeSpec.PART_ADMISSION) {
+                    tags.add(DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES);
+                }
+                break;
+            case ONE_V_TWO_TWO_LEGAL_REP:
+                if (caseData.getShowConditionFlags().contains(DefendantResponseShowTag.CAN_ANSWER_RESPONDENT_1)
+                    && caseData.getRespondent1ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION) {
+                    tags.add(ONLY_RESPONDENT_1_DISPUTES);
+                } else if (caseData.getShowConditionFlags().contains(DefendantResponseShowTag.CAN_ANSWER_RESPONDENT_2)
+                    && caseData.getRespondent2ClaimResponseTypeForSpec() == RespondentResponseTypeSpec.PART_ADMISSION) {
+                    tags.add(DefendantResponseShowTag.ONLY_RESPONDENT_2_DISPUTES);
+                }
+                break;
+            default:
+                throw new UnsupportedOperationException(UNKNOWN_MP_SCENARIO);
+        }
+        return tags;
     }
 
     private Optional<DefendantResponseShowTag> fullDefenceAndPaidLess(
@@ -268,79 +357,17 @@ public class HandleDefendAllClaim implements CaseTask {
         BigDecimal claimedAmount,
         DefendantResponseShowTag ifDisputing,
         DefendantResponseShowTag ifPaidLess) {
-
-        if (isFullDefence(responseType)) {
-            return getDefenceOrPaidLessTag(fullDefenceRoute, responseDetails, claimedAmount, ifDisputing, ifPaidLess);
-        }
-
-        return Optional.empty();
-    }
-
-    private boolean isFullDefence(RespondentResponseTypeSpec responseType) {
-        return FULL_DEFENCE == responseType;
-    }
-
-    private Optional<DefendantResponseShowTag> getDefenceOrPaidLessTag(
-        String fullDefenceRoute, RespondToClaim responseDetails, BigDecimal claimedAmount,
-        DefendantResponseShowTag ifDisputing, DefendantResponseShowTag ifPaidLess) {
-
-        if (isDisputing(fullDefenceRoute)) {
-            return Optional.ofNullable(ifDisputing);
-        } else if (isPaidLess(responseDetails, claimedAmount)) {
-            return Optional.ofNullable(ifPaidLess);
-        }
-
-        return Optional.empty();
-    }
-
-    private boolean isDisputing(String fullDefenceRoute) {
-        return DISPUTES_THE_CLAIM.equals(fullDefenceRoute);
-    }
-
-    private boolean isPaidLess(RespondToClaim responseDetails, BigDecimal claimedAmount) {
-        return Optional.ofNullable(responseDetails)
-            .map(RespondToClaim::getHowMuchWasPaid)
-            .map(MonetaryConversions::penniesToPounds)
-            .map(wasPaid -> wasPaid.compareTo(claimedAmount) < 0)
-            .orElse(false);
-    }
-
-    private void populateRespondentResponseTypeSpecPaidStatus(CaseData caseData, CaseData.CaseDataBuilder<?, ?> updated) {
-        log.info("Populating respondent response type spec paid status for caseData: {}", caseData);
-        setRespondentPaidStatus(caseData, caseData.getDefenceRouteRequired(), caseData.getRespondToClaim(),
-                                updated::respondent1ClaimResponsePaymentAdmissionForSpec);
-
-        if (YES.equals(caseData.getIsRespondent2())) {
-            setRespondentPaidStatus(caseData, caseData.getDefenceRouteRequired2(), caseData.getRespondToClaim2(),
-                                    updated::respondent1ClaimResponsePaymentAdmissionForSpec);
-        }
-    }
-
-    private void setRespondentPaidStatus(CaseData caseData, String defenceRoute, RespondToClaim respondToClaim,
-                                         java.util.function.Consumer<RespondentResponseTypeSpecPaidStatus> updatePaidStatus) {
-        if (SpecJourneyConstantLRSpec.HAS_PAID_THE_AMOUNT_CLAIMED.equals(defenceRoute)
-            && respondToClaim.getHowMuchWasPaid() != null) {
-            int comparison = respondToClaim.getHowMuchWasPaid()
-                .compareTo(new BigDecimal(MonetaryConversions.poundsToPennies(caseData.getTotalClaimAmount())));
-            if (comparison < 0) {
-                updatePaidStatus.accept(RespondentResponseTypeSpecPaidStatus.PAID_LESS_THAN_CLAIMED_AMOUNT);
-            } else {
-                updatePaidStatus.accept(RespondentResponseTypeSpecPaidStatus.PAID_FULL_OR_MORE_THAN_CLAIMED_AMOUNT);
+        if (FULL_DEFENCE == responseType) {
+            if (DISPUTES_THE_CLAIM.equals(fullDefenceRoute)) {
+                return Optional.ofNullable(ifDisputing);
+            } else if (Optional.ofNullable(responseDetails)
+                .map(RespondToClaim::getHowMuchWasPaid)
+                .map(MonetaryConversions::penniesToPounds)
+                .map(wasPaid1 -> wasPaid1.compareTo(claimedAmount) < 0)
+                .orElse(false)) {
+                return Optional.ofNullable(ifPaidLess);
             }
-        } else {
-            updatePaidStatus.accept(RespondentResponseTypeSpecPaidStatus.DID_NOT_PAY);
         }
-    }
-
-    private AllocatedTrack getAllocatedTrack(CaseData caseData) {
-        log.info("Getting allocated track for caseData: {}", caseData);
-        return AllocatedTrack.getAllocatedTrack(caseData.getTotalClaimAmount(), null, null, toggleService, caseData);
-    }
-
-    private CallbackResponse buildCallbackResponse(CaseData.CaseDataBuilder<?, ?> updatedCaseData) {
-        log.info("Building callback response for updated case data");
-        return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(updatedCaseData.build().toMap(objectMapper))
-            .build();
+        return Optional.empty();
     }
 }
