+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/respondtoclaimspeccallbackhandlertaskstests/HandleRespondentResponseTypeForSpec.java
@@ -21,22 +21,13 @@ public class HandleRespondentResponseTypeForSpec implements CaseTask {
     private final ObjectMapper objectMapper;
 
     public CallbackResponse execute(CallbackParams callbackParams) {
-        log.info("Executing HandleRespondentResponseTypeForSpec with callbackParams: {}", callbackParams);
-
         CaseData caseData = callbackParams.getCaseData();
-        log.debug("Initial CaseData: {}", caseData);
-
         if (caseData.getRespondent1ClaimResponseTypeForSpec() != RespondentResponseTypeSpec.FULL_ADMISSION
             || caseData.getRespondent2ClaimResponseTypeForSpec() != RespondentResponseTypeSpec.FULL_ADMISSION) {
-            log.info("Setting specDefenceFullAdmittedRequired to NO as one or both respondents did not fully admit.");
             caseData = caseData.toBuilder().specDefenceFullAdmittedRequired(NO).build();
         }
-
-        CallbackResponse response = AboutToStartOrSubmitCallbackResponse.builder()
+        return AboutToStartOrSubmitCallbackResponse.builder()
             .data(caseData.toMap(objectMapper))
             .build();
-
-        log.debug("CallbackResponse: {}", response);
-        return response;
     }
 }
