+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/respondtoclaimspeccallbackhandlertaskstests/ValidateLengthOfUnemployment.java
@@ -18,27 +18,19 @@ import java.util.List;
 public class ValidateLengthOfUnemployment implements CaseTask {
 
     public CallbackResponse execute(CallbackParams callbackParams) {
-        log.info("Executing ValidateLengthOfUnemployment task");
         CaseData caseData = callbackParams.getCaseData();
         List<String> errors = new ArrayList<>();
 
-        if (caseData.getRespondToClaimAdmitPartUnemployedLRspec() != null
-            && caseData.getRespondToClaimAdmitPartUnemployedLRspec().getLengthOfUnemployment() != null) {
-
-            log.debug("Length of unemployment data found for respondent");
-            if (caseData.getRespondToClaimAdmitPartUnemployedLRspec().getLengthOfUnemployment()
-                .getNumberOfYearsInUnemployment().contains(".")
-                || caseData.getRespondToClaimAdmitPartUnemployedLRspec().getLengthOfUnemployment()
-                .getNumberOfMonthsInUnemployment().contains(".")) {
-
-                log.warn("Length of time unemployed contains decimal values");
-                errors.add("Length of time unemployed must be a whole number, for example, 10.");
+        if (caseData.getRespondToClaimAdmitPartUnemployedLRspec() != null) {
+            var lengthOfUnemployment = caseData.getRespondToClaimAdmitPartUnemployedLRspec().getLengthOfUnemployment();
+            if (lengthOfUnemployment != null) {
+                if (lengthOfUnemployment.getNumberOfYearsInUnemployment().contains(".")
+                    || lengthOfUnemployment.getNumberOfMonthsInUnemployment().contains(".")) {
+                    errors.add("Length of time unemployed must be a whole number, for example, 10.");
+                }
             }
-        } else {
-            log.info("No length of unemployment data found for respondent");
         }
 
-        log.debug("Errors found: {}", errors);
         return AboutToStartOrSubmitCallbackResponse.builder()
             .errors(errors)
             .build();
