+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/respondtoclaimspeccallbackhandlertaskstests/ValidateMediationUnavailableDates.java
@@ -21,23 +21,16 @@ import static uk.gov.hmcts.reform.civil.utils.MediationUnavailableDatesUtils.che
 public class ValidateMediationUnavailableDates implements CaseTask {
 
     public CallbackResponse execute(CallbackParams callbackParams) {
-        log.info("Executing ValidateMediationUnavailableDates task");
         CaseData caseData = callbackParams.getCaseData();
         List<String> errors = new ArrayList<>();
-
-        if (caseData.getResp1MediationAvailability() != null
-            && YES.equals(caseData.getResp1MediationAvailability().getIsMediationUnavailablityExists())) {
-            log.info("Respondent 1 has mediation unavailability dates");
+        if ((caseData.getResp1MediationAvailability() != null
+            && YES.equals(caseData.getResp1MediationAvailability().getIsMediationUnavailablityExists()))) {
             checkUnavailable(errors, caseData.getResp1MediationAvailability().getUnavailableDatesForMediation());
         } else if (caseData.getResp2MediationAvailability() != null
             && YES.equals(caseData.getResp2MediationAvailability().getIsMediationUnavailablityExists())) {
-            log.info("Respondent 2 has mediation unavailability dates");
             checkUnavailable(errors, caseData.getResp2MediationAvailability().getUnavailableDatesForMediation());
-        } else {
-            log.info("No mediation unavailability dates found for respondents");
         }
 
-        log.debug("Errors found: {}", errors);
         return AboutToStartOrSubmitCallbackResponse.builder()
             .errors(errors)
             .build();
