+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/respondtoclaimspeccallbackhandlertaskstests/ValidateRespondentWitnesses.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
-import uk.gov.hmcts.reform.civil.enums.CaseRole;
 import uk.gov.hmcts.reform.civil.enums.MultiPartyScenario;
 import uk.gov.hmcts.reform.civil.handler.callback.user.task.CaseTask;
 import uk.gov.hmcts.reform.civil.model.CaseData;
@@ -29,73 +28,34 @@ public class ValidateRespondentWitnesses implements CaseTask, WitnessesValidator
     private final RespondToClaimSpecUtils respondToClaimSpecUtils;
 
     public CallbackResponse execute(CallbackParams callbackParams) {
-        log.info("Executing ValidateRespondentWitnesses task with callbackParams: {}", callbackParams);
         CaseData caseData = callbackParams.getCaseData();
-        MultiPartyScenario scenario = MultiPartyScenario.getMultiPartyScenario(caseData);
-
-        if (!ONE_V_ONE.equals(scenario)) {
-            log.info("Handling multi-party scenario for caseData: {}", caseData);
-            return handleMultiPartyScenario(callbackParams, caseData);
+        if (!ONE_V_ONE.equals(MultiPartyScenario.getMultiPartyScenario(caseData))) {
+            if (respondToClaimSpecUtils.isSolicitorRepresentsOnlyOneOfRespondents(callbackParams, RESPONDENTSOLICITORONE)) {
+                return validateR1Witnesses(caseData);
+            } else if (respondToClaimSpecUtils.isSolicitorRepresentsOnlyOneOfRespondents(callbackParams, RESPONDENTSOLICITORTWO)) {
+                return validateWitnesses(callbackParams.getCaseData().getRespondent2DQ());
+            } else if (respondent2HasSameLegalRep(caseData)
+                && caseData.getRespondentResponseIsSame() != null && caseData.getRespondentResponseIsSame() == NO
+                && caseData.getRespondent2DQ() != null && caseData.getRespondent2DQ().getRespondent2DQWitnesses() != null) {
+                return validateWitnesses(callbackParams.getCaseData().getRespondent2DQ());
+            }
         }
-        log.info("Validating R1 witnesses for caseData: {}", caseData);
         return validateR1Witnesses(caseData);
-    }
-
-    private CallbackResponse handleMultiPartyScenario(CallbackParams callbackParams, CaseData caseData) {
-        if (isSolicitorRepresentsOnlyOneOfRespondents(callbackParams, RESPONDENTSOLICITORONE)) {
-            log.info("Solicitor represents only one of the respondents (RESPONDENTSOLICITORONE)");
-            return validateR1Witnesses(caseData);
-        } else if (isSolicitorRepresentsOnlyOneOfRespondents(callbackParams, RESPONDENTSOLICITORTWO)) {
-            log.info("Solicitor represents only one of the respondents (RESPONDENTSOLICITORTWO)");
-            return validateWitnesses(caseData.getRespondent2DQ());
-        } else if (isRespondent2HasDifferentLegalRep(caseData)) {
-            log.info("Respondent 2 has a different legal representative");
-            return validateWitnesses(caseData.getRespondent2DQ());
-        }
-        log.info("Defaulting to validate R1 witnesses");
-        return validateR1Witnesses(caseData);
-    }
-
-    private boolean isSolicitorRepresentsOnlyOneOfRespondents(CallbackParams callbackParams, CaseRole caseRole) {
-        boolean result = respondToClaimSpecUtils.isSolicitorRepresentsOnlyOneOfRespondents(callbackParams, caseRole);
-        log.info("isSolicitorRepresentsOnlyOneOfRespondents for caseRole {}: {}", caseRole, result);
-        return result;
-    }
-
-    private boolean isRespondent2HasDifferentLegalRep(CaseData caseData) {
-        boolean result = isSameLegalRep(caseData) && isRespondentResponseNotSame(caseData) && isRespondent2DQWitnesses(caseData);
-        log.info("isRespondent2HasDifferentLegalRep: {}", result);
-        return result;
-    }
-
-    private boolean isSameLegalRep(CaseData caseData) {
-        boolean result = respondToClaimSpecUtils.isRespondent2HasSameLegalRep(caseData);
-        log.info("isSameLegalRep: {}", result);
-        return result;
-    }
-
-    private boolean isRespondentResponseNotSame(CaseData caseData) {
-        boolean result = caseData.getRespondentResponseIsSame() != null && caseData.getRespondentResponseIsSame() == NO;
-        log.info("isRespondentResponseNotSame: {}", result);
-        return result;
-    }
-
-    private boolean isRespondent2DQWitnesses(CaseData caseData) {
-        boolean result = caseData.getRespondent2DQ() != null && caseData.getRespondent2DQ().getRespondent2DQWitnesses() != null;
-        log.info("isRespondent2DQWitnesses: {}", result);
-        return result;
     }
 
     private CallbackResponse validateR1Witnesses(CaseData caseData) {
         List<String> errors = new ArrayList<>();
         if (caseData.getRespondent1DQWitnessesRequiredSpec() == YES
             && caseData.getRespondent1DQWitnessesDetailsSpec() == null) {
-            log.info("Witness details required for Respondent 1");
             errors.add("Witness details required");
         }
-        log.info("Validation errors for R1 witnesses: {}", errors);
         return AboutToStartOrSubmitCallbackResponse.builder()
             .errors(errors)
             .build();
+    }
+
+    private boolean respondent2HasSameLegalRep(CaseData caseData) {
+        return caseData.getRespondent2SameLegalRepresentative() != null
+            && caseData.getRespondent2SameLegalRepresentative() == YES;
     }
 }
