+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/respondtoclaimspeccallbackhandlertaskstests/HandleDefendAllClaimTest.java
@@ -55,8 +55,7 @@ class HandleDefendAllClaimTest {
         handleDefendAllClaim = new HandleDefendAllClaim(
             objectMapper,
             toggleService,
-            paymentDateValidator,
-            respondToClaimSpecUtilsDisputeDetails
+            paymentDateValidator
         );
     }
